+++ src/Pinky/Brainfuck/Optimizer.hs
@@ -7,7 +7,7 @@ where
 import Pinky.Brainfuck.Language
 import Pinky.Brainfuck.Optimizer.Internal
 import Pinky.Brainfuck.Machine
-import Pinky.Brainfuck.Tape
+import Pinky.Brainfuck.Tape.Vector
 
-interpretOptimized :: (BfOptCell c, BrainfuckMachine m c) => Bf -> m ()
-interpretOptimized = evalBfTape . runBfEffects . runBfOptState . parseRaw
+interpretOptimized :: (BfOptCell c, BrainfuckMachine c m) => Bf -> m ()
+interpretOptimized = evalBfOptTape . runBfEffects . runBfOptState . parseRaw