+++ src/Pinky/Brainfuck/Tape.hs
@@ -1,52 +1,16 @@
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE StandaloneDeriving #-}
-{-# LANGUAGE TemplateHaskell #-}
-{-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE FunctionalDependencies #-}
 
 module Pinky.Brainfuck.Tape
-  ( BfState (..),
-    BrainfuckTape,
-    BfCell (..),
+  ( MonadBfTape (..),
     BfError (..),
-    truncateTape,
-    moveHeadLeft,
-    moveHeadRight,
-    moveHead,
-    readCell,
-    readCellOffset,
-    modifyCell,
-    modifyCellOffset,
-    setCell,
-    setCellOffset,
-    runBfTape,
-    evalBfTape,
   )
 where
 
-import Control.Exception
-import Control.Monad.State.Strict
-import Data.Char
-import Data.Functor
-import Data.Word
-import Type.Reflection
-import Pinky.Brainfuck.Machine
-
-data BfState t = BfState
-  { _bfTapeLeft :: ![t],
-    _bfHead :: !t,
-    _bfTapeRight :: ![t]
-  }
-  deriving (Show, Eq)
-
-initTape :: BfCell t => BfState t
-initTape = BfState [] 0 (repeat 0)
-
--- | Trim the right side of the tape.
-truncateTape :: Int -> BfState c -> BfState c
-truncateTape size (BfState l h r) = BfState l h (take size r)
+import Control.Exception (Exception (..))
+import Control.Monad (replicateM_)
+import Control.Monad.Trans (MonadTrans)
+import Pinky.Brainfuck.Machine (BfCell)
+import Type.Reflection (Typeable)
 
 data BfError
   = BadToCharConv
@@ -61,92 +25,58 @@ instance Exception BfError where
   displayException (NegativeTapeIndex i) =
     "tried to modify negative tape index " ++ show i
 
--- | Monad for describing Brainfuck tape manipulations
+-- | Monad class for describing Brainfuck tape manipulations
 --
--- Contains an infinite type (single-directional) of cells storing type t, all
+-- Contains an infinite (single-directional) tape of cells storing type c, all
 -- initialized to 0 and a tape head, starting at the left-most cell of the
--- tape. There is some minimal debugging capability which debugs the current
--- head and tape contents.
+-- tape.
 --
 -- On runtime errors, such as overflowing off the left side of the tape, a
 -- BfError is thrown. Note that errors are thrown "lazily"; for example, a tape
--- overflow error will not be thrown as long as no reads or writes are
+-- overflow error may not be thrown as long as no reads or writes are
 -- performed off the edge of the tape.
 --
 -- While the primitive Brainfuck instructions are supported, there are also
 -- some operations which can operate on parts of the tape not under the head
 -- for optimizing implementations to use.
-newtype BrainfuckTape t m a = BrainfuckTape
-  { unBfTape :: StateT (BfState t) m a
-  }
-  deriving (Functor, Applicative, Monad)
+class (BfCell c, Monad m) => MonadBfTape c m | m -> c where
+  moveHeadLeft :: m ()
+  moveHeadLeft = moveHead (-1)
 
-instance MonadTrans (BrainfuckTape t) where
-  lift = BrainfuckTape . lift
+  moveHeadRight :: m ()
+  moveHeadRight = moveHead 1
 
-moveHeadLeft :: Monad m => BrainfuckTape t m ()
-moveHeadLeft = BrainfuckTape $ do
-  BfState l h r <- get
-  put $ BfState (tail l) (head l) (h : r)
+  moveHead :: Int -> m ()
+  moveHead n
+    | n > 0 = replicateM_ n moveHeadRight
+    | n < 0 = replicateM_ (- n) moveHeadLeft
+    | otherwise = pure ()
 
-moveHeadRight :: Monad m => BrainfuckTape t m ()
-moveHeadRight = BrainfuckTape $ do
-  BfState l h r <- get
-  put $ BfState (h : l) (head r) (tail r)
+  readCell :: m c
+  readCell = readCellOffset 0
 
-moveHead :: Monad m => Int -> BrainfuckTape t m ()
-moveHead n
-  | n > 0 = replicateM_ n moveHeadRight
-  | n < 0 = replicateM_ (- n) moveHeadLeft
-  | otherwise = pure ()
+  readCellOffset :: Int -> m c
+  readCellOffset n = do
+    moveHead n
+    v <- readCell
+    moveHead (- n)
+    pure v
 
-readCell :: (Monad m, BfCell t) => BrainfuckTape t m t
-readCell = BrainfuckTape $ gets _bfHead
+  -- TODO Make writeCell in terms of modifyCell instead of the other way around?
 
-readCellOffset :: (Monad m, BfCell t) => Int -> BrainfuckTape t m t
-readCellOffset offset = do
-  moveHead offset
-  cell <- readCell
-  moveHead (- offset)
-  return cell
+  writeCell :: c -> m ()
+  writeCell = writeCellOffset 0
 
-modifyCell :: (Monad m, BfCell t) => (t -> t) -> BrainfuckTape t m ()
-modifyCell f = BrainfuckTape $ do
-  tape <- get
-  put $ tape {_bfHead = f (_bfHead tape)}
+  writeCellOffset :: Int -> c -> m ()
+  writeCellOffset n v = do
+    moveHead n
+    writeCell v
+    moveHead (- n)
 
-modifyCellOffset ::
-  (Monad m, BfCell t) => Int -> (t -> t) -> BrainfuckTape t m ()
-modifyCellOffset offset f = do
-  moveHead offset
-  modifyCell f
-  moveHead (- offset)
+  modifyCell :: (c -> c) -> m ()
+  modifyCell = modifyCellOffset 0
 
-setCell :: (Monad m, BfCell t) => t -> BrainfuckTape t m ()
-setCell = modifyCell . const
-
-setCellOffset :: (Monad m, BfCell t) => Int -> t -> BrainfuckTape t m ()
-setCellOffset off = modifyCellOffset off . const
-
-runBfTape :: (Monad m, BfCell t) => BrainfuckTape t m a -> m (a, BfState t)
-runBfTape bf = runStateT (unBfTape bf) initTape
-
-evalBfTape :: (Monad m, BfCell t) => BrainfuckTape t m a -> m a
-evalBfTape = fmap fst . runBfTape
-
-{-}
-debugBf :: (Monad m, BfCell t) => BrainfuckTape t IO ()
-debugBf = do
-  BfState tape pos <- getState
-  let len = M.length tape
-  liftIO $ do
-    putStrLn ""
-    putStrLn $ "head = " ++ show pos
-    forM_
-      [0 .. len - 1]
-      ( \i -> do
-          val <- M.read tape i
-          putStr $ show val ++ ","
-      )
-    putStrLn ""
-    -}
+  modifyCellOffset :: Int -> (c -> c) -> m ()
+  modifyCellOffset n f = do
+    v <- readCellOffset n
+    writeCellOffset n (f v)